Update API update course

+++ KnowledgeSharingApi/KnowledgeSharingApi.Domain/Models/ApiRequestModels/UpdateUserItemModels/UpdateCourseModel.cs
@@ -17,7 +17,7 @@ namespace KnowledgeSharingApi.Domains.Models.ApiRequestModels.UpdateUserItemMode
 
         public string? Introduction { get; set; }
 
-        public int? Fee { get; set; }
+        public decimal? Fee { get; set; }
 
         public int? EstimateTimeInMinutes { get; set; }
     }
